fix + rename package, format eclipse
